refactor(ui-logic): extract game logic into dedicated hooks across 15 games
Each game component now delegates all state management, timers, canvas loops,
and business logic to a co-located `use<Game>.ts` hook. Components are pure
renderers that destructure the hook's return value and map it to JSX.

Affected games: answer-cannon, letter-slingshot, syllable-drums, letter-grow,
letter-bubble-shooter, letter-slicer, word-fishing, cooking-game,
spot-the-difference, dress-up, team-picker, timer, and SpeedBurstTimer
(shared component now uses useSpeedBurst from hooks/shared/game-modes/).
