+++ gamesformykids/app/games/dress-up/DressUpClient.tsx
@@ -1,122 +1,13 @@
 'use client';
-
-import { useState, useCallback, useEffect, useRef } from 'react';
+import { useDressUpGame } from './useDressUpGame';
 import { speakHebrew } from '@/lib/utils/speech/speaker';
 
-type Zone = 'head' | 'neck' | 'torso' | 'hands' | 'legs' | 'feet';
-type Category = 'daily' | 'seasonal' | 'professions';
-type Phase = 'menu' | 'playing' | 'result';
-
-interface ClothingItem {
-  name: string;
-  hebrew: string;
-  prompt: string;
-  emoji: string;
-  zone: Zone;
-  category: Category;
-}
-
-const CLOTHING: ClothingItem[] = [
-  // Daily
-  { name: 'hat',        hebrew: 'כּוֹבַע',          prompt: 'שים כובע!',          emoji: '🧢', zone: 'head',   category: 'daily' },
-  { name: 'shirt',      hebrew: 'חוּלְצָה',          prompt: 'שים חולצה!',         emoji: '👕', zone: 'torso',  category: 'daily' },
-  { name: 'dress',      hebrew: 'שִׂמְלָה',          prompt: 'שים שמלה!',          emoji: '👗', zone: 'torso',  category: 'daily' },
-  { name: 'pants',      hebrew: 'מִכְנָסַיִם',        prompt: 'שים מכנסיים!',       emoji: '👖', zone: 'legs',   category: 'daily' },
-  { name: 'shoes',      hebrew: 'נַעֲלַיִם',          prompt: 'שים נעליים!',        emoji: '👟', zone: 'feet',   category: 'daily' },
-  { name: 'socks',      hebrew: 'גַּרְבַּיִם',        prompt: 'שים גרביים!',        emoji: '🧦', zone: 'feet',   category: 'daily' },
-  // Seasonal
-  { name: 'coat',       hebrew: 'מְעִיל',            prompt: 'שים מעיל!',          emoji: '🧥', zone: 'torso',  category: 'seasonal' },
-  { name: 'scarf',      hebrew: 'צָעִיף',            prompt: 'שים צעיף!',          emoji: '🧣', zone: 'neck',   category: 'seasonal' },
-  { name: 'gloves',     hebrew: 'כְּפָפוֹת',          prompt: 'שים כפפות!',         emoji: '🧤', zone: 'hands',  category: 'seasonal' },
-  { name: 'boots',      hebrew: 'מַגָּפַיִם',          prompt: 'שים מגפיים!',        emoji: '👢', zone: 'feet',   category: 'seasonal' },
-  { name: 'sunglasses', hebrew: 'מִשְׁקְפֵי שֶׁמֶשׁ', prompt: 'שים משקפי שמש!',    emoji: '🕶️', zone: 'head',   category: 'seasonal' },
-  { name: 'umbrella',   hebrew: 'מִטְרִיָּה',          prompt: 'שים מטרייה!',        emoji: '☂️', zone: 'hands',  category: 'seasonal' },
-  // Professions
-  { name: 'hard-hat',   hebrew: 'קַסְדָּה',           prompt: 'שים קסדה!',          emoji: '⛑️', zone: 'head',   category: 'professions' },
-  { name: 'graduation', hebrew: 'כּוֹבַע סִיּוּם',    prompt: 'שים כובע סיום!',     emoji: '🎓', zone: 'head',   category: 'professions' },
-  { name: 'lab-coat',   hebrew: 'חֲלוּק לָבָן',       prompt: 'שים חלוק לבן!',      emoji: '🥼', zone: 'torso',  category: 'professions' },
-  { name: 'backpack',   hebrew: 'תִּיק גַּב',          prompt: 'שים תיק גב!',        emoji: '🎒', zone: 'hands',  category: 'professions' },
-  { name: 'stethoscope',hebrew: 'סְטֶתוֹסְקוֹפ',      prompt: 'שים סטטוסקופ!',      emoji: '🩺', zone: 'neck',   category: 'professions' },
-];
-
-const ZONE_ORDER: Zone[] = ['head', 'neck', 'torso', 'hands', 'legs', 'feet'];
-const ZONE_PLACEHOLDER: Record<Zone, string> = {
-  head: '❓', neck: '❓', torso: '❓', hands: '❓', legs: '❓', feet: '❓',
-};
-const ZONE_LABEL: Record<Zone, string> = {
-  head: 'ראש', neck: 'צוואר', torso: 'גוף', hands: 'ידיים', legs: 'רגליים', feet: 'נעליים',
-};
-
-const QUESTIONS_PER_GAME = 8;
-const CATEGORY_LABELS: Record<Category, string> = {
-  daily: '👕 לבוש יומיומי',
-  seasonal: '🌦️ לבוש עונתי',
-  professions: '🎓 לבוש מקצועי',
-};
-
-function shuffle<T>(arr: T[]): T[] {
-  return [...arr].sort(() => Math.random() - 0.5);
-}
-
-function buildChoices(correct: ClothingItem, pool: ClothingItem[]): ClothingItem[] {
-  const others = shuffle(pool.filter(c => c.name !== correct.name)).slice(0, 3);
-  return shuffle([correct, ...others]);
-}
-
 export default function DressUpClient() {
-  const [phase, setPhase] = useState<Phase>('menu');
-  const [questions, setQuestions] = useState<ClothingItem[]>([]);
-  const [qIdx, setQIdx] = useState(0);
-  const [choices, setChoices] = useState<ClothingItem[]>([]);
-  const [dressed, setDressed] = useState<Partial<Record<Zone, ClothingItem>>>({});
-  const [score, setScore] = useState(0);
-  const [feedback, setFeedback] = useState<'correct' | 'wrong' | null>(null);
-  const timerRef = useRef<ReturnType<typeof setTimeout> | null>(null);
-
-  const startGame = useCallback(() => {
-    const qs = shuffle(CLOTHING).slice(0, QUESTIONS_PER_GAME);
-    setQuestions(qs);
-    setQIdx(0);
-    setScore(0);
-    setDressed({});
-    setFeedback(null);
-    const first = qs[0]!;
-    setChoices(buildChoices(first, CLOTHING));
-    setPhase('playing');
-    setTimeout(() => speakHebrew(first.prompt), 400);
-  }, []);
-
-  const current = questions[qIdx];
-
-  const selectAnswer = useCallback((item: ClothingItem) => {
-    if (feedback || !current) return;
-    if (timerRef.current) clearTimeout(timerRef.current);
-
-    if (item.name === current.name) {
-      setFeedback('correct');
-      setScore(s => s + 1);
-      setDressed(d => ({ ...d, [current.zone]: current }));
-      speakHebrew(`כן! ${current.hebrew}!`);
-      timerRef.current = setTimeout(() => {
-        const next = qIdx + 1;
-        if (next >= questions.length) {
-          setPhase('result');
-        } else {
-          const nextQ = questions[next]!;
-          setQIdx(next);
-          setChoices(buildChoices(nextQ, CLOTHING));
-          setFeedback(null);
-          setTimeout(() => speakHebrew(nextQ.prompt), 300);
-        }
-      }, 1300);
-    } else {
-      setFeedback('wrong');
-      speakHebrew('לא נכון, נסה שוב!');
-      timerRef.current = setTimeout(() => setFeedback(null), 900);
-    }
-  }, [current, feedback, qIdx, questions]);
-
-  useEffect(() => () => { if (timerRef.current) clearTimeout(timerRef.current); }, []);
+  const {
+    phase, qIdx, choices, dressed, score, feedback,
+    startGame, selectAnswer, current,
+    ZONE_ORDER, ZONE_PLACEHOLDER, ZONE_LABEL, CATEGORY_LABELS, QUESTIONS_PER_GAME,
+  } = useDressUpGame();
 
   if (phase === 'menu') {
     return (
@@ -175,7 +66,6 @@ export default function DressUpClient() {
   // Playing phase
   return (
     <div className="min-h-screen bg-gradient-to-br from-pink-100 via-purple-50 to-purple-200 flex flex-col items-center p-4" dir="rtl">
-      {/* Header */}
       <div className="w-full max-w-md flex justify-between items-center mb-4 pt-2">
         <span className="bg-white rounded-full px-4 py-1 text-purple-700 font-bold text-lg shadow">
           {qIdx + 1} / {QUESTIONS_PER_GAME}
@@ -186,10 +76,8 @@ export default function DressUpClient() {
         </span>
       </div>
 
-      {/* Character display */}
       <div className="bg-white rounded-3xl shadow-xl p-4 mb-4 w-full max-w-md">
         <div className="flex justify-center gap-6">
-          {/* Body zones */}
           <div className="flex flex-col gap-2 items-center">
             {ZONE_ORDER.map(zone => (
               <div key={zone} className="flex items-center gap-2">
@@ -206,14 +94,12 @@ export default function DressUpClient() {
               </div>
             ))}
           </div>
-          {/* Character figure */}
           <div className="flex flex-col items-center justify-center text-6xl leading-none">
             <span>🧑</span>
           </div>
         </div>
       </div>
 
-      {/* Question */}
       <div className={`bg-white rounded-2xl shadow-lg p-4 mb-4 w-full max-w-md text-center border-4 transition-colors ${
         feedback === 'correct' ? 'border-green-400 bg-green-50' :
         feedback === 'wrong' ? 'border-red-300 bg-red-50' :
@@ -234,7 +120,6 @@ export default function DressUpClient() {
         )}
       </div>
 
-      {/* Choices grid */}
       <div className="grid grid-cols-2 gap-3 w-full max-w-md">
         {choices.map(item => (
           <button
