+++ gamesformykids/app/games/spot-the-difference/SpotClient.tsx
@@ -1,198 +1,11 @@
 'use client';
-import { useState, useCallback, useRef, useEffect } from 'react';
-import { speakHebrew } from '@/lib/utils/speech/speaker';
-
-interface Diff {
-  index: number;
-  altEmoji: string;
-  hint: string;
-}
-
-interface Scene {
-  name: string;
-  emoji: string;
-  bg: string;
-  cols: number;
-  items: string[];
-  diffs: Diff[];
-}
-
-const SCENES: Scene[] = [
-  {
-    name: 'חיות', emoji: '🐾', bg: 'from-green-100 to-emerald-200',
-    cols: 5,
-    items: [
-      '🐶','🐱','🐭','🐹','🐰',
-      '🦊','🐻','🐼','🐨','🐯',
-      '🦁','🐮','🐷','🐸','🐵',
-      '🐔','🐧','🐦','🐤','🦜',
-    ],
-    diffs: [
-      { index: 2,  altEmoji: '🐹', hint: 'הכלב שינה' },
-      { index: 7,  altEmoji: '🐨', hint: 'הפנדה שינה' },
-      { index: 11, altEmoji: '🦙', hint: 'הפרה שינתה' },
-      { index: 16, altEmoji: '🦚', hint: 'הפינגווין שינה' },
-      { index: 18, altEmoji: '🐦', hint: 'האפרוח שינה' },
-    ],
-  },
-  {
-    name: 'מטבח', emoji: '🍴', bg: 'from-orange-100 to-yellow-200',
-    cols: 5,
-    items: [
-      '🍕','🍔','🌮','🌯','🥗',
-      '🍱','🍛','🍲','🥘','🥣',
-      '🥞','🧇','🥓','🥚','🍳',
-      '🥩','🍗','🍖','🌽','🥦',
-    ],
-    diffs: [
-      { index: 0,  altEmoji: '🍰', hint: 'הפיצה שינתה' },
-      { index: 5,  altEmoji: '🍣', hint: 'הקופסה שינתה' },
-      { index: 10, altEmoji: '🥐', hint: 'הלביבות שינו' },
-      { index: 15, altEmoji: '🍖', hint: 'הבשר שינה' },
-      { index: 18, altEmoji: '🥒', hint: 'התירס שינה' },
-    ],
-  },
-  {
-    name: 'גן', emoji: '🌳', bg: 'from-lime-100 to-green-200',
-    cols: 5,
-    items: [
-      '🌳','🌲','🌴','🌵','🌾',
-      '🍀','🌿','🌱','🌸','🌺',
-      '🌻','🌼','💐','🌷','🌹',
-      '🏡','🚗','🚕','🐕','🐈',
-    ],
-    diffs: [
-      { index: 1,  altEmoji: '🌺', hint: 'עץ שינה' },
-      { index: 6,  altEmoji: '🌸', hint: 'הצמח שינה' },
-      { index: 12, altEmoji: '🌻', hint: 'הזר שינה' },
-      { index: 17, altEmoji: '🚙', hint: 'המכונית שינתה' },
-      { index: 19, altEmoji: '🐱', hint: 'החתול שינה' },
-    ],
-  },
-  {
-    name: 'ים', emoji: '🌊', bg: 'from-blue-100 to-cyan-200',
-    cols: 5,
-    items: [
-      '🐠','🐟','🐡','🦈','🐙',
-      '🦑','🦐','🦞','🦀','🐚',
-      '🪸','🐚','🐬','🐳','🐋',
-      '🦭','🐊','🐢','🌊','💧',
-    ],
-    diffs: [
-      { index: 0,  altEmoji: '🐡', hint: 'הדג שינה' },
-      { index: 4,  altEmoji: '🦑', hint: 'התמנון שינה' },
-      { index: 9,  altEmoji: '🪸', hint: 'הצדפה שינתה' },
-      { index: 13, altEmoji: '🐋', hint: 'הדולפין שינה' },
-      { index: 17, altEmoji: '🐟', hint: 'הצב שינה' },
-    ],
-  },
-  {
-    name: 'שמים', emoji: '🌟', bg: 'from-indigo-100 to-purple-200',
-    cols: 5,
-    items: [
-      '☀️','🌤️','⛅','🌥️','☁️',
-      '🌦️','🌧️','⛈️','🌩️','🌨️',
-      '❄️','🌬️','🌈','🌙','⭐',
-      '✨','💫','🌟','🌠','🌌',
-    ],
-    diffs: [
-      { index: 0,  altEmoji: '🌝', hint: 'השמש שינתה' },
-      { index: 4,  altEmoji: '🌩️', hint: 'הענן שינה' },
-      { index: 9,  altEmoji: '☃️', hint: 'השלג שינה' },
-      { index: 14, altEmoji: '🌑', hint: 'הכוכב שינה' },
-      { index: 17, altEmoji: '💥', hint: 'הכוכב שינה' },
-    ],
-  },
-];
-
-const TIMER_START = 90;
-const DIFF_POINTS = 20;
-
-type Phase = 'menu' | 'playing' | 'result';
+import { useSpotGame, SCENES } from './useSpotGame';
 
 export default function SpotClient() {
-  const [phase, setPhase]         = useState<Phase>('menu');
-  const [sceneIdx, setSceneIdx]   = useState(0);
-  const [found, setFound]         = useState<Set<number>>(new Set());
-  const [timeLeft, setTimeLeft]   = useState(TIMER_START);
-  const [score, setScore]         = useState(0);
-  const [wrongIdx, setWrongIdx]   = useState<number | null>(null);
-  const [shakePanel, setShake]    = useState(false);
-  const timerRef = useRef<ReturnType<typeof setInterval> | null>(null);
-  const flashRef = useRef<ReturnType<typeof setTimeout> | null>(null);
-
-  const stopTimer = useCallback(() => {
-    if (timerRef.current) clearInterval(timerRef.current);
-  }, []);
-
-  const startScene = useCallback((idx: number, initialScore = 0) => {
-    setSceneIdx(idx);
-    setFound(new Set());
-    setTimeLeft(TIMER_START);
-    setScore(initialScore);
-    setWrongIdx(null);
-    setPhase('playing');
-    const s = SCENES[idx]!;
-    setTimeout(() => speakHebrew(`מצא 5 הבדלים ב${s.name}!`), 300);
-  }, []);
-
-  const startGame = useCallback(() => {
-    const idx = Math.floor(Math.random() * SCENES.length);
-    startScene(idx, 0);
-  }, [startScene]);
-
-  // Countdown timer
-  useEffect(() => {
-    if (phase !== 'playing') return;
-    timerRef.current = setInterval(() => {
-      setTimeLeft(t => {
-        if (t <= 1) {
-          stopTimer();
-          setPhase('result');
-          return 0;
-        }
-        return t - 1;
-      });
-    }, 1000);
-    return stopTimer;
-  }, [phase, stopTimer]);
-
-  const handleTap = useCallback((index: number, _fromAltPanel: boolean) => {
-    if (phase !== 'playing') return;
-    const scene = SCENES[sceneIdx]!;
-    const diff = scene.diffs.find(d => d.index === index);
-
-    if (diff && !found.has(index)) {
-      const newFound = new Set(found);
-      newFound.add(index);
-      setFound(newFound);
-      const newScore = score + DIFF_POINTS;
-      setScore(newScore);
-      speakHebrew(`כן! מצאת! ${diff.hint}`);
-
-      if (newFound.size >= scene.diffs.length) {
-        stopTimer();
-        setTimeout(() => {
-          setPhase('result');
-          speakHebrew(`מדהים! מצאת את כל ה-5 הבדלים!`);
-        }, 600);
-      }
-    } else if (!diff || found.has(index)) {
-      setWrongIdx(index);
-      setShake(true);
-      speakHebrew('לא הבדל כאן!');
-      if (flashRef.current) clearTimeout(flashRef.current);
-      flashRef.current = setTimeout(() => {
-        setWrongIdx(null);
-        setShake(false);
-      }, 500);
-    }
-  }, [phase, sceneIdx, found, score, stopTimer]);
-
-  useEffect(() => () => {
-    stopTimer();
-    if (flashRef.current) clearTimeout(flashRef.current);
-  }, [stopTimer]);
+  const {
+    phase, sceneIdx, found, timeLeft, score, wrongIdx, shakePanel,
+    setPhase, startGame, startScene, handleTap, stopTimer, scene,
+  } = useSpotGame();
 
   if (phase === 'menu') {
     return (
@@ -264,9 +77,6 @@ export default function SpotClient() {
   }
 
   // Playing
-  const scene = SCENES[sceneIdx]!;
-
-  // Build panel B items
   const itemsA = scene.items;
   const itemsB = scene.items.map((e, i) => {
     const d = scene.diffs.find(d => d.index === i);
@@ -285,16 +95,16 @@ export default function SpotClient() {
         style={{ gridTemplateColumns: `repeat(${scene.cols}, 1fr)` }}
       >
         {items.map((emoji, i) => {
-          const isFound   = scene.diffs.some(d => d.index === i) && found.has(i);
-          const isWrong   = wrongIdx === i;
+          const isFound = scene.diffs.some(d => d.index === i) && found.has(i);
+          const isWrong = wrongIdx === i;
           return (
             <button
               key={i}
-              onClick={() => handleTap(i, panel === 'B')}
+              onClick={() => handleTap(i)}
               className={`
                 aspect-square text-xl sm:text-2xl flex items-center justify-center rounded-lg transition-all
-                ${isFound  ? 'bg-green-300 ring-2 ring-green-500 scale-105' : ''}
-                ${isWrong  ? 'bg-red-200 scale-95' : ''}
+                ${isFound ? 'bg-green-300 ring-2 ring-green-500 scale-105' : ''}
+                ${isWrong ? 'bg-red-200 scale-95' : ''}
                 ${!isFound && !isWrong ? 'hover:bg-purple-100 active:scale-95' : ''}
               `}
             >
@@ -309,7 +119,6 @@ export default function SpotClient() {
 
   return (
     <div className={`min-h-screen flex flex-col bg-gradient-to-br ${scene.bg} p-3`}>
-      {/* HUD */}
       <div className="flex items-center justify-between mb-2">
         <div className="flex items-center gap-2">
           <span className="text-2xl">{scene.emoji}</span>
@@ -319,7 +128,6 @@ export default function SpotClient() {
         <div className="bg-yellow-400 text-gray-800 font-bold px-3 py-1 rounded-full text-sm">{score}</div>
       </div>
 
-      {/* Progress */}
       <div className="flex items-center gap-2 mb-2">
         <span className="text-sm text-purple-700">הבדלים:</span>
         {scene.diffs.map((d, i) => (
@@ -333,7 +141,6 @@ export default function SpotClient() {
         <span className="text-sm text-purple-700 mr-1">{found.size}/5</span>
       </div>
 
-      {/* Two panels */}
       <div className="flex flex-col sm:flex-row gap-2 flex-1">
         {renderPanel(itemsA, 'A')}
         {renderPanel(itemsB, 'B')}
