+++ gamesformykids/app/games/timer/TimerClient.tsx
@@ -1,100 +1,22 @@
 'use client';
-
-import { useState, useEffect, useCallback, useRef } from 'react';
-import { speakHebrew } from '@/lib/utils/speech/speaker';
-
-const PRESETS = [
-  { label: '30 שנ׳', seconds: 30 },
-  { label: '1 דקה', seconds: 60 },
-  { label: '2 דק׳', seconds: 120 },
-  { label: '3 דק׳', seconds: 180 },
-  { label: '5 דק׳', seconds: 300 },
-  { label: '10 דק׳', seconds: 600 },
-];
-
-function formatTime(s: number) {
-  const m = Math.floor(s / 60);
-  const sec = s % 60;
-  return `${String(m).padStart(2, '0')}:${String(sec).padStart(2, '0')}`;
-}
+import { useState } from 'react';
+import { useTimer, PRESETS, formatTime } from './useTimer';
 
 export default function TimerClient() {
-  const [total, setTotal] = useState(60);
-  const [remaining, setRemaining] = useState(60);
-  const [running, setRunning] = useState(false);
-  const [done, setDone] = useState(false);
+  const {
+    total, remaining, running, done,
+    start, pause, reset, selectPreset, applyCustom, toggleFullscreen,
+    pct, warning, progressColor,
+  } = useTimer();
+
   const [customMin, setCustomMin] = useState('');
   const [customSec, setCustomSec] = useState('');
-  const intervalRef = useRef<ReturnType<typeof setInterval> | null>(null);
 
-  const clear = useCallback(() => {
-    if (intervalRef.current) clearInterval(intervalRef.current);
-    intervalRef.current = null;
-  }, []);
-
-  useEffect(() => {
-    return clear;
-  }, [clear]);
-
-  const start = useCallback(() => {
-    if (remaining <= 0) return;
-    setDone(false);
-    setRunning(true);
-    intervalRef.current = setInterval(() => {
-      setRemaining((r) => {
-        if (r <= 1) {
-          clearInterval(intervalRef.current!);
-          intervalRef.current = null;
-          setRunning(false);
-          setDone(true);
-          speakHebrew('הזמן נגמר!');
-          return 0;
-        }
-        return r - 1;
-      });
-    }, 1000);
-  }, [remaining]);
-
-  const pause = useCallback(() => {
-    clear();
-    setRunning(false);
-  }, [clear]);
-
-  const reset = useCallback(() => {
-    clear();
-    setRunning(false);
-    setDone(false);
-    setRemaining(total);
-  }, [clear, total]);
-
-  const selectPreset = useCallback((seconds: number) => {
-    clear();
-    setRunning(false);
-    setDone(false);
-    setTotal(seconds);
-    setRemaining(seconds);
+  const handleSelectPreset = (seconds: number) => {
+    selectPreset(seconds);
     setCustomMin('');
     setCustomSec('');
-  }, [clear]);
-
-  const applyCustom = useCallback(() => {
-    const m = parseInt(customMin || '0', 10);
-    const s = parseInt(customSec || '0', 10);
-    const t = m * 60 + s;
-    if (t > 0) selectPreset(t);
-  }, [customMin, customSec, selectPreset]);
-
-  const toggleFullscreen = useCallback(() => {
-    if (!document.fullscreenElement) {
-      document.documentElement.requestFullscreen().catch(() => {});
-    } else {
-      document.exitFullscreen().catch(() => {});
-    }
-  }, []);
-
-  const pct = total > 0 ? (remaining / total) * 100 : 0;
-  const warning = remaining <= 10 && remaining > 0 && !done;
-  const progressColor = done ? '#22c55e' : warning ? '#ef4444' : '#60a5fa';
+  };
 
   return (
     <div
@@ -179,7 +101,7 @@ export default function TimerClient() {
         {PRESETS.map(({ label, seconds }) => (
           <button
             key={seconds}
-            onClick={() => selectPreset(seconds)}
+            onClick={() => handleSelectPreset(seconds)}
             className={`px-4 py-2 rounded-xl font-bold text-sm transition-[transform,background-color] active:scale-95 ${
               total === seconds && !customMin && !customSec
                 ? 'bg-white text-indigo-700 shadow-lg'
@@ -213,7 +135,7 @@ export default function TimerClient() {
           className="w-14 text-center bg-white/20 text-white font-bold rounded-xl px-2 py-1.5 placeholder:text-white/40 focus:outline-none focus:ring-2 focus:ring-white/50"
         />
         <button
-          onClick={applyCustom}
+          onClick={() => applyCustom(customMin, customSec)}
           className="px-3 py-1.5 bg-blue-400 hover:bg-blue-300 active:scale-95 transition-transform text-white font-bold rounded-xl text-sm"
         >
           הגדר
